backend/worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "celery>=5.6.3",
    "elasticsearch>=9.3.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "lxml>=6.1.0",
    "minio>=7.2.20",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.13.3",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "redis>=7.4.0",
    "sentence-transformers>=5.4.1",
]
